bot/pyproject.toml
[tool.poetry]
name = "telegram-nginx-docker-webhook"
version = "0.1.0"
description = ""
authors = ["cnwv <115669384+cnwv@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
aiogram = "^3.3.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
